+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 * FlatLaf UI theme
 * Add theme setting
 * Make mod list clip long mod names instead of scrolling horizontally
+* Add UI to switch between play, out-jar, and package
 
 #### v3.30.3 ####
 * Fix package information being lost for patched classes (Alchyr)